Remove explicit VACUUM ANALYZE
The referenced docs have changed. Up until version 14 of PostgreSQL they
stated this:
```
We recommend that active production databases be vacuumed frequently (at
least nightly), in order to remove dead rows. After adding or
deleting a large number of rows, it might be a good idea to issue a
VACUUM ANALYZE command for the affected table. This will update the
system catalogs with the results of all recent changes, and allow the
PostgreSQL query planner to make better choices in planning queries.
```

From version 15 it states:
```
We recommend that all databases be vacuumed regularly in order to remove
dead rows. PostgreSQL includes an “autovacuum” facility which can
automate routine vacuum maintenance. For more information about
automatic and manual vacuuming, see Section 25.1.
```

So if anything one should probably look at tweaking the autovacuum
config over doing things explicitly. I have seen that even with the
explicit call for VACUUM ANALYZE the "sites" table will still grow with
bloat over time. This is probably a result of continously replacing the
million rows in the sites table when replacing a list.

The proper solution to the bloat problem is probably to use partitioned
sites tables where instead of doing DELETE you just drop the related
partition.