Start using a partitioned sites table
The previous setup with a single sites table where we would insert an
updated set of sites and then delete (via ON DELETE CASCADE) the
existing site rows would end up creating ever increasing postgres bloat.

I noticed this where a small VPS had slowly been filling up its disk with
database bloat until it was full even if the code did run explicit
VACUUM ANALYZE when the sites table was modified. A "VACUUM FULL sites"
did give back the storage but that is not optimal since it locks the
table etc.

Now we have a partitoned sites table where cleaning up an old list means
we first delete the old list as before, but then also drop the related
sites partition table entirely. This should stop bloat from adding up
over time.

This meant dropping the id BIGSERIAL PRIMARY KEY column on the sites
table as uniqueness is only enforced within partitions which initially
resulted in a "pq: unique constraint on partitioned table must include all
partitioning columns" error on initial attempts. Having an ID column
that is not globally unique to the table does not seem helpful so remove
it in favor of using the combination of list_id and site name as PRIMARY
KEY instead.

While here add GRANT statements to initDB() to make recreation of the
tables actually allow tlapid to function afterwards.

Finally add a check that stops tldbwriter if it detects that you are
running an non-partitioned sites table with a hint to clean it up. It
would probably be nice to actually handle data migration instead.